requirements.txt
GitPython==3.1.37
PyYAML==6.0.1
Requests==2.32.3
setuptools==75.8.0
websockets==10.4
pydantic>=1.10.0
